Copy Experiments (#352)
* clone endpoint

* attributes are now cloned correctly

* finished clone endpoint

* experiment duplication in UI & management of multiple experiment versions

* remove unnecessary comment

* fix min.get

* duplication button in experiment page

* improved review page

* reference to old version during review

* fix teamscale

* use dropdown menu

* fix

* make dropdowm optional

* fix bug on file deletion

* fix teamscale

* resolve merge conflict

* clean

---------

+++ server/api/files/[id].delete.ts
@@ -19,14 +19,27 @@ export default defineEventHandler(async (event) => {
 
   await authorize(event, fileAbilities.delete, file)
 
+  const referencedExperiments = await prisma.experiment.count({
+    where: {
+      previewImageId: file.id,
+    },
+  })
+
+  if (referencedExperiments) {
+    return setResponseStatus(event, 204)
+  }
+
   const runtimeConfig = useRuntimeConfig()
   const filePath = `${runtimeConfig.fileMount}/${file.path}`
 
-  myDeleteFile(filePath)
-
-  await prisma.file.delete({
-    where,
-  })
+  try {
+    await prisma.file.delete({
+      where,
+    })
+    myDeleteFile(filePath)
+  } catch {
+    // don't delete file because there are multiple references to it
+  }
 
   return setResponseStatus(event, 204)
 })
